requirements.txt
psycopg2==2.9.1
dataclasses==0.6
great_expectations==0.13.35
numpy==1.21.2
pandas==1.3.3
python-dotenv==0.19.0
riotwatcher==3.1.4
ruamel.base==1.0.0
